Remove daemon executable
